Fix more headers, mote libpthread to libc

+++ sysroot/usr/include/stdint.h
@@ -33,7 +33,7 @@
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned long uint64_t;
 
 // intN_t
 typedef signed char int8_t;
